Update yotpo sync method to save overall score in product
